example

+++ src/main/java/com/booking/domain/OwnerReview.java
@@ -5,7 +5,7 @@ import java.util.Date;
 public class OwnerReview extends Review {
     private Long OwnerId;
 
-    public OwnerReview(int reviewId, Date reviewDate, String description, float rating, Long ownerId) {
+    public OwnerReview(Long reviewId, Date reviewDate, String description, float rating, Long ownerId) {
         super(reviewId, reviewDate, description, rating);
         OwnerId = ownerId;
     }
